Send generic file and bugfix
Now is possible to send every type of file. If the file is not an image, the MIME type will be 'text/plain'.

Fixed the bug that would crash the program when someone would send the same image in a Telegram chat.

Moved the test classes to test package.

Renamed the class from "bottelegram" to "telegrambot"